backend/requirements.txt
pydantic==2.5.2
firebase-admin==6.2.0
cloudflare-python==1.0.1
python-jose==3.3.0
